refactoring drop through block process and each scene factor staging

+++ sdl_week27/include/GameConfig.hpp
@@ -10,7 +10,7 @@ namespace GameConfig{
     // 得点用係数
     static inline  constexpr double SCORE_RATE = 100.0;
     // GameOver時のペナルティ
-    static inline constexpr double SCORE_PENALTY = -5000.0;
+    static inline constexpr int SCORE_PENALTY = -5000.0;
 
     // デバッグ用の制御用変数
     // fpsなどの情報の制御用
